Add flags to choose which experiments to setup and run

+++ baselines/bbw/experiments/scalability.sh
@@ -8,13 +8,19 @@ WIKITABLES=${BENCHMARK_DIR}wikitables_2013/tables/
 SLEEP=8h
 
 # Tough Tables - Wikidata
-python bbwWrapper.py ${TOUGH_TABLES_WD} toughtables_wd_scalability &
-PID=$1
-sleep ${SLEEP}
-kill -9 ${PID}
+if [[ -d ${TOUGH_TABLES_WD} ]]
+then
+    python bbwWrapper.py ${TOUGH_TABLES_WD} toughtables_wd_scalability &
+    PID=$1
+    sleep ${SLEEP}
+    kill -9 ${PID}
+fi
 
 # Wikitables
-python bbwWrapper.py ${WIKITABLES} wikitables_scalability &
-PID=$1
-sleep ${SLEEP}
-kill -9 ${PID}
+if [[ -d ${WIKITABLES} ]]
+then
+    python bbwWrapper.py ${WIKITABLES} wikitables_scalability &
+    PID=$1
+    sleep ${SLEEP}
+    kill -9 ${PID}
+fi
