+++ baselines/emblookup/experiments/scalability.sh
@@ -10,25 +10,34 @@ DATA_DIR="data/"
 SLEEP=8h
 
 # ToughTables - DBpedia
-mkdir -p /results/emblookup/toughtables_dbp_scalability
+if [[ -d ${TOUGH_TABLES_DBP} ]]
+then
+    mkdir -p /results/emblookup/toughtables_dbp_scalability
 
-python3 main.py "${TOUGH_TABLES_DBP}" "${DATA_DIR}aliases/alias_dbp10-2016.ttl" "${DATA_DIR}index_mappings/kg_index_name_mapping_dbp10-2016.csv" "${DATA_DIR}aliases_processed/aliases_processed_dbp10-2016.csv" -h &
-PID=$1
-sleep ${SLEEP}
-kill -9 ${PID}
+    python3 main.py "${TOUGH_TABLES_DBP}" "${DATA_DIR}aliases/alias_dbp10-2016.ttl" "${DATA_DIR}index_mappings/kg_index_name_mapping_dbp10-2016.csv" "${DATA_DIR}aliases_processed/aliases_processed_dbp10-2016.csv" -h &
+    PID=$1
+    sleep ${SLEEP}
+    kill -9 ${PID}
+fi
 
 # ToughTables - Wikidata
-mkdir -p /results/emblookup/toughtables_wd_scalability
+if [[ -d ${TOUGH_TABLES_WD} ]]
+then
+    mkdir -p /results/emblookup/toughtables_wd_scalability
 
-python3 main.py "${TOUGH_TABLES_WD}" "${DATA_DIR}aliases/alias_wd.ttl" "${DATA_DIR}index_mappings/kg_index_name_mapping_wd.csv" "${DATA_DIR}aliases_processed/aliases_processed_wd.csv" -h &
-PID=$1
-sleep ${SLEEP}
-kill -9 ${PID}
+    python3 main.py "${TOUGH_TABLES_WD}" "${DATA_DIR}aliases/alias_wd.ttl" "${DATA_DIR}index_mappings/kg_index_name_mapping_wd.csv" "${DATA_DIR}aliases_processed/aliases_processed_wd.csv" -h &
+    PID=$1
+    sleep ${SLEEP}
+    kill -9 ${PID}
+fi
 
 # Wikitables
-mkdir -p /results/emblookup/wikitables_dbp_2013_scalability
+if [[ -d ${WIKITABLES} ]]
+then
+    mkdir -p /results/emblookup/wikitables_dbp_2013_scalability
 
-python3 main.py "${WIKITABLES}" "${DATA_DIR}aliases/alias_dbp12-2022.ttl" "${DATA_DIR}index_mappings/kg_index_name_mapping_dbp12-2022.csv" "${DATA_DIR}aliases_processed/aliases_processed_dbp12-2022.csv" -h &
-PID=$1
-sleep ${SLEEP}
-kill -9 ${PID}
+    python3 main.py "${WIKITABLES}" "${DATA_DIR}aliases/alias_dbp12-2022.ttl" "${DATA_DIR}index_mappings/kg_index_name_mapping_dbp12-2022.csv" "${DATA_DIR}aliases_processed/aliases_processed_dbp12-2022.csv" -h &
+    PID=$1
+    sleep ${SLEEP}
+    kill -9 ${PID}
+fi
